requirements.txt
beautifulsoup4==4.14.2
fastapi==0.122.0
httpx==0.28.1
matplotlib==3.10.7
numpy==2.3.5
pandas==2.3.3
pdfplumber==0.11.8
Pillow==12.0.0
playwright==1.56.0
plotly==6.5.0
pydantic==2.12.5
PyPDF2==3.0.1
python-dotenv==1.2.1
Requests==2.32.5
seaborn==0.13.2
uvicorn==0.38.0
